Move `getInput` to `Expectations` in helper package.
